requirements.txt
python-pptx==0.6.21
opencc-python-reimplemented==0.1.7
